update build.xml and pom.xml to include random shape generator Rename shape generator in spatial 3d and remove it from spatial extras update random shape generator to generate GeoPointShapes Update RandomGeoRelationshipTest to cope with GeoPointShape Update Geo3dSpatialContextFactory so it is constructed with a valid default state.
