+++ lucene/spatial3d/src/test/org/apache/lucene/spatial3d/geom/RandomBinaryCodecTest.java
@@ -27,7 +27,7 @@ import org.junit.Test;
 /**
  * Test to check Serialization
  */
-public class RandomBinaryCodecTest extends RandomGeoShapeGenerator{
+public class RandomBinaryCodecTest extends RandomGeo3dShapeGenerator {
 
   @Test
   @Repeat(iterations = 10)
